Programs 2-5
 [Use OpenGL Library to: ]
Q2. Write a program in C to draw a line by using DDA algorithm when starting point, slope and number of iterations are given.
Q3. Write a program in C to draw two lines by using DDA algorithm when a common starting point, their slopes (unequal) and number of iterations are given.
Q4. Write a program in C to draw a line by using Bresenham Line Drawing Algorithm when two points are given.
Q5. Write a program in C to draw two parallel lines when two points are given for the first line and one point as starting point where x−value is same as in the starting point of the first line and y−value is increased by n (a natural number) for the second line by using Bresenham Line Drawing Algorithm.